+++ p2panda-auth/src/lib.rs
@@ -2,3 +2,6 @@
 
 pub mod group;
 pub mod traits;
+mod access;
+
+pub use access::{Access, AccessLevel};
